+++ src/js/integrations/reactnativeerrorhandlers.ts
@@ -3,6 +3,7 @@ import type { EventHint, Integration, SeverityLevel } from '@sentry/types';
 import { addExceptionMechanism, logger } from '@sentry/utils';
 
 import type { ReactNativeClient } from '../client';
+import { createSyntheticError } from '../utils/error';
 import { RN_GLOBAL_OBJ } from '../utils/worldwide';
 
 /** ReactNativeErrorHandlers Options */
@@ -131,6 +132,7 @@ export class ReactNativeErrorHandlers implements Integration {
         getCurrentHub().captureException(error, {
           data: { id },
           originalException: error,
+          syntheticException: createSyntheticError(),
         });
       },
       onHandled: (id: string) => {